+++ src/main/java/com/anthonini/brewer/service/exception/NotPossibleDeleteUserException.java
@@ -1,4 +1,4 @@
-package com.anthonini.brewer.service;
+package com.anthonini.brewer.service.exception;
 
 public class NotPossibleDeleteUserException extends RuntimeException {
